+++ projects/configgen/configgen/emulatorlauncher.py
@@ -145,6 +145,10 @@ def getGenerator(emulator):
         module = __import__("configgen.generators.vvvvvv.vvvvvvGenerator", fromlist=["VvvvvvGenerator"])
         generatorClass = getattr(module, "VvvvvvGenerator")
         return generatorClass()
+    elif emulator == "vpinball":
+        module = __import__("configgen.generators.vpinball.vpinballGenerator", fromlist=["VpinballGenerator"])
+        generatorClass = getattr(module, "VpinballGenerator")
+        return generatorClass()
     else:
         print("Missing generator for {}".format(emulator))
         raise ValueError