+++ projects/configgen/configgen/recalboxFiles.py
@@ -56,6 +56,8 @@ recalboxBins =\
     'vice'        : '/usr/bin/x64',
     'xroar'       : '/usr/bin/xroar',
     'vvvvvv'      : '/usr/bin/vvvvvv/VVVVVV',
+    'vpinball'    : '/usr/bin/vpinball/VPinballX_GL',
+
 }
 
 
@@ -182,3 +184,6 @@ frotzConfig = HOME + '/.config/frotz/frotz.conf'
 
 sdlpopConfigFile = CONF + '/sdlpop/sdlpop.ini'
 sdlpopControllersFile = CONF + '/sdlpop/gamecontrollerdb.txt'
+
+vpinballRootFolder = CONF + '/vpinball'
+vpinballConfigFile = vpinballRootFolder + '/VPinballX-configgen.ini'